refactor: move contents of `app` to `static`
