+++ pages/page/[page].tsx
@@ -4,7 +4,7 @@ import { useContext } from 'react';
 import GameGrid from '../../components/GameGrid';
 import Legend from '../../components/Legend';
 import Overview from '../../components/Overview';
-import { SettingsContext } from '../../src/app/state';
+import { SettingsContext } from '../../src/static/state';
 import { allImages } from '../../src/assets';
 import { Games, paginationSize } from '../../src/static';
 import { paginate, stats } from '../../src/utils/games';
