+++ src/static/icons.tsx
